+++ demos/src/Extensions/InvisibleCharacters/Vue/index.spec.js
@@ -1,0 +1,9 @@
+context('/src/Extensions/InvisibleCharacters/Vue/', () => {
+  before(() => {
+    cy.visit('/src/Extensions/InvisibleCharacters/Vue/')
+  })
+
+  it('should have invisible characters', () => {
+    cy.get('[class*="Tiptap-invisible-character"]').should('exist')
+  })
+})