+++ packages/extension-file-handler/src/index.ts
@@ -1,0 +1,8 @@
+import { FileHandler } from './fileHandler.js'
+
+export * from './FileHandlePlugin.js'
+export * from './types.js'
+
+export { FileHandler }
+
+export default FileHandler