+++ packages/extension-font-family/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 3.0.0-beta.10
+
+### Patch Changes
+
+- @tiptap/extension-text-style@3.0.0-beta.10
+
+## 3.0.0-beta.9
+
+### Patch Changes
+
+- @tiptap/extension-text-style@3.0.0-beta.9
+
 ## 3.0.0-beta.8
 
 ### Patch Changes