+++ packages/extension-mention/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## 3.0.0-beta.10
+
+### Patch Changes
+
+- @tiptap/core@3.0.0-beta.10
+- @tiptap/pm@3.0.0-beta.10
+- @tiptap/suggestion@3.0.0-beta.10
+
+## 3.0.0-beta.9
+
+### Patch Changes
+
+- @tiptap/core@3.0.0-beta.9
+- @tiptap/pm@3.0.0-beta.9
+- @tiptap/suggestion@3.0.0-beta.9
+
 ## 3.0.0-beta.8
 
 ### Patch Changes