+++ packages/react/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 3.0.0-beta.10
+
+### Patch Changes
+
+- @tiptap/core@3.0.0-beta.10
+- @tiptap/pm@3.0.0-beta.10
+
+## 3.0.0-beta.9
+
+### Patch Changes
+
+- ecb0dee: Fixes an issue where ref props will be kept on non-supported react components
+  - @tiptap/core@3.0.0-beta.9
+  - @tiptap/pm@3.0.0-beta.9
+
 ## 3.0.0-beta.8
 
 ### Patch Changes